Switched build system to premake4
- Fixed some warnings in player and savestation
- Removed binaries and normal makefiles

+++ src/objects/savestation.c
@@ -55,7 +55,7 @@ void obj_savestation_update(struct tds_object* ptr) {
 void obj_savestation_draw(struct tds_object* ptr) {
 	struct obj_savestation_data* data = (struct obj_savestation_data*) ptr->object_data;
 
-	float glow_mod = ((sin(data->dt_glow) + 1.0f) / 2.0f) * 0.01f;
+	float glow_mod = ((sinf(data->dt_glow) + 1.0f) / 2.0f) * 0.01f;
 
 	struct tds_render_light lt_glow = {TDS_RENDER_LIGHT_POINT, ptr->x, ptr->y, 0.0f, 0.01f + glow_mod, 0.02f + glow_mod * 2.0f, 10.0f, NULL};
 	tds_render_submit_light(tds_engine_global->render_handle, lt_glow);
